division fix

+++ src/ColorAnalysis.hs
@@ -53,7 +53,7 @@ makegraph col=let
 coloranalysis::[Int]->Maybe Color
 coloranalysis col=let
     level=60
-    points=sum col
+    points=fromIntegral (sum col)
     pctg=[(floor((col!!n)/points*100),n)|n<-[0..63]]
     sortedpctg=reverse(sort pctg)
     maxlist=takeWhile (\x->fst x>level) sortedpctg
